Removed unused code

+++ src/runtime_src/xdp/profile/plugin/aie_trace/x86/aie_trace.h
@@ -24,26 +24,26 @@
 
 namespace xdp {
 
-  class AieTrace_x86Impl : public AieTraceImpl{
+  class AieTrace_x86Impl : public AieTraceImpl {
 
     public:
-      AieTrace_x86Impl(VPDatabase* database, std::shared_ptr<AieTraceMetadata> metadata);
+      AieTrace_x86Impl(VPDatabase* database,
+                       std::shared_ptr<AieTraceMetadata> metadata);
       ~AieTrace_x86Impl() = default;
       virtual void updateDevice();
       virtual void flushAieTileTraceModule();
       bool setMetricsSettings(uint64_t deviceId, void* handle);
       uint64_t checkTraceBufSize(uint64_t size);
-      void pollTimers(uint32_t index, void* handle);
+      void pollTimers(uint64_t index, void* handle);
       void parseMessages(uint8_t* messages);
       module_type getTileType(uint16_t absRow);
       void freeResources();
-      
+
     private:
       xrt::device device;
       xrt::kernel aie_trace_kernel;
-    
   };
 
-}   
+} // namespace xdp
 
 #endif
